Refactor AddressToTokenMap contract and improve comments

+++ contracts/AddressToTokenMap.sol
@@ -1,14 +1,16 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.6;
 
-contract AddressToTokenMap{
+contract AddressToTokenMap {
     address owner;
+
     // token address => Symbol for Symbol retrieval
     mapping(address => string) private addresses;
-    // token address => tokenToUSD pair PriceFeed Address
+
+    // token address => token/USD Chainlink price feed address
     mapping(address => address) private priceFeedMap;
 
-    constructor(){
+    constructor() {
         owner = msg.sender;
     }
 
@@ -16,29 +18,30 @@ contract AddressToTokenMap{
         require(owner == msg.sender, "Not Owner, cannot add mapping");
         _;
     }
-    
-    /* 
-    * @dev : retrieves the symbol or a token's address - used when adding assets
-    *        and in determining if the symbol is ETH for eth vs token transfers
-    * @params : token address 
-    * @returns : symbol string
-    */
+
+    /**
+     * @dev Retrieves the symbol for a token address
+     * @param _key Token address
+     * @return symbol string
+     */
     function getSymbol(address _key) public view returns (string memory) {
         return addresses[_key];
     }
 
-    /* 
-    * @dev : maps token address to symbol
-    * @params : token address , symbol string
-    */
+    /**
+     * @dev Maps token address to symbol
+     * @param _key Token address
+     * @param _value Symbol string
+     */
     function _setAddress(address _key, string memory _value) public onlyOwner {
-        // Avoids updating addresses[_key] if the new value is the same as the current value
+        // Avoid updating if the value is unchanged
         bytes memory valueBytes = bytes(_value);
         bytes memory keyBytes = bytes(addresses[_key]);
+
         if (valueBytes.length != keyBytes.length) {
             addresses[_key] = _value;
         } else {
-            for (uint i = 0; i < valueBytes.length; i++) {
+            for (uint256 i = 0; i < valueBytes.length; i++) {
                 if (valueBytes[i] != keyBytes[i]) {
                     addresses[_key] = _value;
                     break;
@@ -47,31 +50,41 @@ contract AddressToTokenMap{
         }
     }
 
-    /* 
-    * @dev : returns the price feed address for a token's address - chainlink price oracle
-    * @params : token address 
-    * @returns : price feed address
-    */
-    function getPriceFeedMap(address _tokenAddress) public view returns(address) {
+    /**
+     * @dev Returns the Chainlink price feed address for a token
+     * @param _tokenAddress Token address
+     * @return price feed address
+     */
+    function getPriceFeedMap(address _tokenAddress)
+        public
+        view
+        returns (address)
+    {
         return priceFeedMap[_tokenAddress];
     }
 
-    /* 
-    * @dev : maps token address to sepolia/mainnet chainlink price feed address
-    * @params : token address , price feed address
-    */
-    function _setPriceFeedMap(address _tokenAddress, address _pairAddress) public onlyOwner{
+    /**
+     * @dev Maps token address to Chainlink price feed address
+     * @param _tokenAddress Token address
+     * @param _pairAddress Price feed address
+     */
+    function _setPriceFeedMap(
+        address _tokenAddress,
+        address _pairAddress
+    ) public onlyOwner {
         if (priceFeedMap[_tokenAddress] != _pairAddress) {
             priceFeedMap[_tokenAddress] = _pairAddress;
         }
     }
 
-    /* 
-    * @dev : returns true fif the address's symbol is ETH
-    * @params : token address
-    * returns : boolean 
-    */
-    function isETH(address _token) public view returns(bool) {
-        return keccak256(abi.encodePacked(getSymbol(_token))) == keccak256(abi.encodePacked("ETH"));
+    /**
+     * @dev Returns true if the token symbol is ETH
+     * @param _token Token address
+     * @return boolean
+     */
+    function isETH(address _token) public view returns (bool) {
+        return
+            keccak256(abi.encodePacked(getSymbol(_token))) ==
+            keccak256(abi.encodePacked("ETH"));
     }
 }
